all: Add scope

+++ app/src/main/java/com/siwiba/wba/model/Saldo.kt
@@ -1,11 +1,11 @@
 package com.siwiba.wba.model
 
 data class Saldo(
-    val no: Int = 0,
+    val no: Long = 0,
     val keterangan: String = "",
-    val debit: Int = 0,
-    val kredit: Int = 0,
-    val saldo: Int = 0,
+    val debit: Long = 0,
+    val kredit: Long = 0,
+    val saldo: Long = 0,
     val editor: String = "",
     val tanggal: String = ""
 )
